Task_6_Snowflake/requirements.txt
apache-airflow-providers-mongo
pandas
pymongo
apache-airflow-providers-snowflake
apache-airflow-providers-common-sql